backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
Pillow==10.1.0
stripe==7.4.0
django-allauth==0.57.0
dj-rest-auth==5.0.2
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
psycopg2-binary==2.9.9
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4